% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeart_random.R
\name{codeart_random}
\alias{codeart_random}
\title{Create Random Code Art}
\usage{
codeart_random(
  w = 1200,
  h = 1200,
  code_lines = 60,
  min_segments = 8,
  max_segments = 20,
  min_segment_length = 15,
  max_segment_length = 60,
  segment_sep = 20,
  line_break_chance = 0.4,
  indent_size = 50,
  max_indents = 6,
  indent_inc_chance = 0.4,
  indent_dec_chance = 0.3,
  change_chance = 0.4,
  cols = NULL
)
}
\description{
Create Random Code Art
}
\details{
Inspired by [Eric Davidson](https://github.com/erdavids/Simulated-Code).
}
\author{
David Schoch
}
