% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeart_script.R
\name{codeart_script}
\alias{codeart_script}
\title{Create Code Art from R script}
\usage{
codeart_script(path, char_size = 5, segment_sep = 20, indent_size = 50, cols)
}
\description{
Create Code Art from R script
}
\details{
Inspired by [Eric Davidson](https://github.com/erdavids/Simulated-Code).
}
\author{
David Schoch
}
