% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_codeart_rand}
\alias{plot_codeart_rand}
\title{Plot Random Code Art}
\usage{
plot_codeart_rand(df, seg_size = 3, bg_col = "grey25")
}
\arguments{
\item{df}{result from running [codeart_random]}

\item{seg_size}{width of code segments}

\item{bg_col}{background color}
}
\description{
Plot Random Code Art
}
\author{
David Schoch
}
