% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_codeart_script}
\alias{plot_codeart_script}
\title{Plot Code Art generated from an R script}
\usage{
plot_codeart_script(
  df,
  seg_size = 3,
  bg_col = "grey25",
  pipe_size = 8,
  caption = "",
  title = "",
  title_size = 62,
  caption_size = 32,
  font = "Arial"
)
}
\arguments{
\item{df}{result from running [codeart_script]}

\item{seg_size}{width of code segments}

\item{bg_col}{background color}

\item{pipe_size}{size of pipes}

\item{caption}{caption}

\item{title}{title}

\item{title_size}{size of title}

\item{caption_size}{size of caption}

\item{font}{font family}
}
\description{
Plot Code Art generated from an R script
}
\author{
David Schoch
}
